♻️ Update latitude and longitude data types to float
The latitude and longitude properties in both MobileFoodFacility and MobileFoodFacilitiesResponse have been modified from int type to float. This change has been done to achieve more precise location data. Additionally, optimized GetAllMobileFoodFacilitiesQueryHandler to handle empty list cases, and adjusted API endpoint in FoodTruck.Api.http for testing purpose.

+++ src/FoodTruck.Contracts/MobileFoodFacilities/MobileFoodFacilitiesResponse.cs
@@ -9,5 +9,5 @@ public record MobileFoodFacilitiesResponse(
     string Address,
     string? BlockLot,
     string? FoodItems,
-    int Longitude,
-    int Latitude);
+    float Longitude,
+    float Latitude);
